go concurrency: add rate limit example
