+++ doc/go_concurrency/code/44_select_break.go
