+++ doc/go_concurrency/code/45_close_channel.go
