+++ doc/go_concurrency/code/46_blocking_defer.go
