+++ doc/go_concurrency/code/47_buffered_channel_copy.go
