+++ doc/go_concurrency/code/50_select_default.go
