+++ doc/go_concurrency/code/53_blocking_mutex.go
