+++ doc/go_concurrency/code/54_empty_buffered_channel.go
