+++ doc/go_concurrency/code/55_close_two.go
