+++ doc/go_concurrency/code/56_atomic_defer.go
