+++ doc/go_concurrency/code/57_channel_capacity.go
